added user-model and usersRouter
