+++ users/users-model.js
@@ -1,0 +1,18 @@
+const db = require("../data/dbConfig");
+
+module.exports = {
+  get,
+  add,
+  getBy,
+};
+
+function get() {
+  return db("users");
+}
+
+function add(user) {
+  return db("users").insert(user);
+}
+function getBy(filter) {
+  return db("users").where(filter).orderBy("id");
+}
